debt: server: extract ws-handlers, http-handlers, banner + anomaly-manager bootstrap
server.js: 1102 → 595 lines.

- server/lib/ws-command-handlers.js: handleWsCommand + the override /
  relay handlers it dispatches to. Owns overrideTtlTimer.
- server/lib/http-handlers.js: createHandlers factory returning the
  HTTP API handler set (health, version, history, events, device-
  config, sensor-discovery, push/*). Takes db, authMiddleware,
  broadcastToWebSockets as deps — server.js constructs handlers in
  initServices' finish() once db is resolved.
- server/lib/banner.js: getNetworkAddress + printBanner. Pure.
- server/lib/anomaly-manager.js: adds bootstrap() helper that
  applies the watchdog_events schema, creates the history backend,
  and calls init() — collapses ~35 lines of initAnomalyManager in
  server.js down to a single call.
- Minor comment trim in the PWA/public route blocks.

npm test:unit 782/782 pass. knip clean.
