+++ server/server.js
@@ -20,8 +20,10 @@ const sensorConfig = require('./lib/sensor-config');
 const sensorDiscovery = require('./lib/sensor-discovery');
 const push = require('./lib/push');
 const anomalyManager = require('./lib/anomaly-manager');
-const { createHistory: createWatchdogHistory } = require('./lib/watchdog-history');
 const { createScriptMonitor } = require('./lib/script-monitor');
+const { handleWsCommand } = require('./lib/ws-command-handlers');
+const { createHandlers, readBody, jsonResponse } = require('./lib/http-handlers');
+const { getNetworkAddress, printBanner } = require('./lib/banner');
 
 const log = createLogger('server');
 const PORT = parseInt(process.env.PORT || process.argv[2] || '3000', 10);
@@ -88,42 +90,6 @@ function serveFile(filePath, urlPath, res) {
   });
 }
 
-// ── Health endpoint ──
-
-function checkVpn(callback) {
-  if (!VPN_CHECK_HOST) { callback('unknown'); return; }
-  var parts = VPN_CHECK_HOST.split(':');
-  var host = parts[0];
-  var port = parseInt(parts[1] || '80', 10);
-  var sock = net.createConnection({ host: host, port: port, timeout: 2000 });
-  sock.on('connect', function () { sock.destroy(); callback('connected'); });
-  sock.on('error', function () { sock.destroy(); callback('disconnected'); });
-  sock.on('timeout', function () { sock.destroy(); callback('disconnected'); });
-}
-
-function handleHealth(req, res) {
-  checkVpn(function (vpnStatus) {
-    var status = vpnStatus === 'disconnected' ? 'degraded' : 'ok';
-    res.writeHead(200, { 'Content-Type': 'application/json' });
-    res.end(JSON.stringify({
-      status: status,
-      vpn: vpnStatus,
-      mqtt: mqttBridge.getConnectionStatus(),
-      timestamp: new Date().toISOString(),
-    }));
-  });
-}
-
-// ── Version endpoint ──
-// Returns the git commit hash for client-side change detection.
-// GIT_COMMIT is baked into the Docker image at build time (see Dockerfile).
-
-var APP_VERSION = process.env.GIT_COMMIT || 'unknown';
-
-function handleVersion(req, res) {
-  res.writeHead(200, { 'Content-Type': 'application/json' });
-  res.end(JSON.stringify({ hash: APP_VERSION }));
-}
 
 // ── Auth middleware ──
 
@@ -133,221 +99,9 @@ if (AUTH_ENABLED) {
   log.info('auth enabled', { rpId: process.env.RPID || 'localhost' });
 }
 
-function readBody(req, callback) {
-  var body = '';
-  req.on('data', function (chunk) { body += chunk; });
-  req.on('end', function () { callback(body); });
-}
-
-function jsonResponse(res, statusCode, data) {
-  res.writeHead(statusCode, { 'Content-Type': 'application/json' });
-  res.end(JSON.stringify(data));
-}
-
-// ── History API ──
-
 var db = null;
 var scriptMonitor = null;
-
-function handleHistoryApi(req, res) {
-  if (!db) {
-    jsonResponse(res, 503, { error: 'Database not available' });
-    return;
-  }
-
-  var parsed = new URL(req.url, 'http://localhost');
-  var range = parsed.searchParams.get('range') || '6h';
-  var sensor = parsed.searchParams.get('sensor') || null;
-
-  db.getHistory(range, sensor, function (err, points) {
-    if (err) {
-      log.error('history query failed', { error: err.message });
-      jsonResponse(res, 500, { error: 'Query failed' });
-      return;
-    }
-    db.getEvents(range, 'mode', function (evErr, events) {
-      if (evErr) {
-        log.error('events query failed', { error: evErr.message });
-        events = [];
-      }
-      jsonResponse(res, 200, { range: range, points: points, events: events });
-    });
-  });
-}
-
-// Paginated state_events feed for the System Logs UI. Supports newest-first
-// cursor pagination so the client can lazy-load older entries on scroll.
-//   GET /api/events?type=mode&limit=10&before=<unix_ms>
-//   -> { events: [...], hasMore: bool }
-function handleEventsApi(req, res) {
-  if (!db) {
-    jsonResponse(res, 503, { error: 'Database not available' });
-    return;
-  }
-
-  var parsed = new URL(req.url, 'http://localhost');
-  var type = parsed.searchParams.get('type') || 'mode';
-  var limit = parseInt(parsed.searchParams.get('limit'), 10) || 10;
-  var beforeRaw = parsed.searchParams.get('before');
-  var before = beforeRaw ? parseInt(beforeRaw, 10) : null;
-  if (beforeRaw && (Number.isNaN(before) || before <= 0)) {
-    jsonResponse(res, 400, { error: 'Invalid `before` cursor' });
-    return;
-  }
-
-  db.getEventsPaginated(type, limit, before, function (err, result) {
-    if (err) {
-      log.error('events query failed', { error: err.message });
-      jsonResponse(res, 500, { error: 'Query failed' });
-      return;
-    }
-    jsonResponse(res, 200, result);
-  });
-}
-
-// ── Device Config API ──
-
-function handleDeviceConfigApi(req, res, urlPath, body) {
-  if (req.method === 'GET') {
-    deviceConfig.handleGet(req, res);
-    return;
-  }
-
-  if (req.method === 'PUT') {
-    // PUT requires admin role
-    if (AUTH_ENABLED && !authMiddleware.requireAdmin(req, res)) return;
-    deviceConfig.handlePut(req, res, body, function (config) {
-      // Publish to MQTT for instant push to device
-      mqttBridge.publishConfig(config);
-      // Mirror the latest snapshot into anomaly-manager so wb-clear
-      // actions from the Mode Enablement UI are reflected in WS
-      // broadcasts and subsequent getState() calls.
-      anomalyManager.updateSnapshot(config);
-      // Push a watchdog-state broadcast so UIs re-render the mode
-      // enablement / watchdog cards live.
-      broadcastToWebSockets({
-        type: 'watchdog-state',
-        pending: anomalyManager.getPending(),
-        watchdogs: require('../shelly/watchdogs-meta.js').WATCHDOGS,
-        snapshot: { we: config.we || {}, wz: config.wz || {}, wb: config.wb || {} }
-      });
-    });
-    return;
-  }
-
-  jsonResponse(res, 405, { error: 'Method not allowed' });
-}
-
-// ── Sensor Discovery API ──
-
-function handleSensorDiscovery(req, res, body) {
-  var parsed;
-  try { parsed = JSON.parse(body); } catch (e) {
-    jsonResponse(res, 400, { error: 'Invalid JSON body' });
-    return;
-  }
-  if (!parsed.hosts || !Array.isArray(parsed.hosts)) {
-    jsonResponse(res, 400, { error: 'Missing hosts array' });
-    return;
-  }
-  // Direct HTTP scan of each hub in parallel. Bypasses the Pro 4PM + MQTT
-  // path because it was slow and produced opaque timeouts on a single hub
-  // failure. See server/lib/sensor-discovery.js.
-  var options = {};
-  if (parsed.skipTemp) options.skipTemp = true;
-  sensorDiscovery.discoverSensors(parsed.hosts, options).then(function (result) {
-    jsonResponse(res, 200, result);
-  }).catch(function (err) {
-    jsonResponse(res, 500, { error: err.message || String(err) });
-  });
-}
-
-// ── Push subscription handlers ──
-
-function handlePushSubscribe(req, res, body) {
-  var parsed;
-  try { parsed = JSON.parse(body); } catch (e) {
-    jsonResponse(res, 400, { error: 'Invalid JSON' });
-    return;
-  }
-  if (!parsed.subscription || !parsed.subscription.endpoint || !parsed.subscription.keys) {
-    jsonResponse(res, 400, { error: 'Missing subscription object (endpoint + keys)' });
-    return;
-  }
-  var categories = Array.isArray(parsed.categories) ? parsed.categories : [];
-  push.addSubscription(parsed.subscription, categories, function (err) {
-    if (err) {
-      jsonResponse(res, 500, { error: 'Failed to save subscription' });
-      return;
-    }
-    jsonResponse(res, 200, { ok: true });
-  });
-}
-
-function handlePushUnsubscribe(req, res, body) {
-  var parsed;
-  try { parsed = JSON.parse(body); } catch (e) {
-    jsonResponse(res, 400, { error: 'Invalid JSON' });
-    return;
-  }
-  if (!parsed.endpoint) {
-    jsonResponse(res, 400, { error: 'Missing endpoint' });
-    return;
-  }
-  push.removeSubscription(parsed.endpoint, function (err) {
-    if (err) {
-      jsonResponse(res, 500, { error: 'Failed to remove subscription' });
-      return;
-    }
-    jsonResponse(res, 200, { ok: true });
-  });
-}
-
-function handlePushGetSubscription(req, res, body) {
-  var parsed;
-  try { parsed = JSON.parse(body); } catch (e) {
-    jsonResponse(res, 400, { error: 'Invalid JSON' });
-    return;
-  }
-  if (!parsed.endpoint) {
-    jsonResponse(res, 400, { error: 'Missing endpoint' });
-    return;
-  }
-  var sub = push.getSubscription(parsed.endpoint);
-  if (!sub) {
-    jsonResponse(res, 200, { subscribed: false, categories: [] });
-    return;
-  }
-  jsonResponse(res, 200, { subscribed: true, categories: sub.categories });
-}
-
-// Send a mock notification of the given category to the caller's
-// subscription. Bypasses rate limiting and category filtering so the
-// user can preview how each notification type renders on their device.
-function handlePushTest(req, res, body) {
-  var parsed;
-  try { parsed = JSON.parse(body); } catch (e) {
-    jsonResponse(res, 400, { error: 'Invalid JSON' });
-    return;
-  }
-  if (!parsed.endpoint || !parsed.category) {
-    jsonResponse(res, 400, { error: 'Missing endpoint or category' });
-    return;
-  }
-  var payload = push.buildMockPayload(parsed.category);
-  if (!payload) {
-    jsonResponse(res, 400, { error: 'Unknown category: ' + parsed.category });
-    return;
-  }
-  push.sendTestToEndpoint(parsed.endpoint, payload, function (err) {
-    if (err) {
-      var status = err.message === 'Subscription not found' ? 404 : 500;
-      jsonResponse(res, status, { error: err.message });
-      return;
-    }
-    jsonResponse(res, 200, { ok: true });
-  });
-}
+var handlers = null;
 
 // ── HTTP route detection (for OTel span naming) ──
 
@@ -382,13 +136,13 @@ var server = http.createServer(function (req, res) {
 
   // Health — always accessible
   if (urlPath === '/health') {
-    handleHealth(req, res);
+    handlers.handleHealth(req, res);
     return;
   }
 
   // Version — always accessible (no sensitive data)
   if (urlPath === '/version') {
-    handleVersion(req, res);
+    handlers.handleVersion(req, res);
     return;
   }
 
@@ -400,33 +154,17 @@ var server = http.createServer(function (req, res) {
     return;
   }
 
-  // /public/* — accessible without auth. Everything the login page and
-  // any other unauthenticated view needs (HTML, CSS, JS, fonts) lives
-  // under playground/public/ so this single prefix check covers them
-  // all. Keep the directory public-by-convention: do not put anything
-  // sensitive there.
+  // /public/* — login page + shared assets (HTML/CSS/JS/fonts) that
+  // must load before auth. Public-by-convention: no secrets in there.
   if (urlPath.startsWith('/public/')) {
     serveStatic(req, res);
     return;
   }
 
-  // PWA static assets — accessible without auth.
-  // Chrome fetches the manifest and icons with `credentials: 'omit'` by
-  // default, so they must be reachable without a session cookie or the
-  // install flow breaks (Chrome receives the 302 to the login page and
-  // tries to parse the HTML as JSON, failing installability). These files contain
-  // no sensitive data — just brand metadata, icons, and the service
-  // worker script — so exempting them from auth is standard practice
-  // for auth-gated PWAs (Slack, Discord, Notion, etc.).
-  //
-  // Paths covered:
-  //   /sw.js                  — service worker script
-  //   /manifest.webmanifest   — PWA manifest
-  //   /assets/icon-*.png      — app icons listed in the manifest
-  //   /assets/badge-*.png     — Android status-bar silhouette used as
-  //                             `badge` in showNotification()
-  //   /assets/notif-*.png     — per-category notification icons sent by
-  //                             the SW via event.data.icon
+  // PWA static assets — Chrome fetches the manifest + icons with
+  // `credentials: 'omit'` so they must be reachable without a session
+  // cookie or the install flow breaks. Safe: they contain no sensitive
+  // data — brand metadata, icons, service-worker script.
   if (
     urlPath === '/sw.js' ||
     urlPath === '/manifest.webmanifest' ||
@@ -490,7 +228,7 @@ var server = http.createServer(function (req, res) {
   // Authenticated routes
   if (urlPath === '/api/device-config') {
     readBody(req, function (body) {
-      handleDeviceConfigApi(req, res, urlPath, body);
+      handlers.handleDeviceConfigApi(req, res, urlPath, body);
     });
   } else if (urlPath === '/api/sensor-config') {
     readBody(req, function (body) {
@@ -530,31 +268,31 @@ var server = http.createServer(function (req, res) {
     if (req.method === 'POST') {
       if (!isAdminOrReject()) return;
       readBody(req, function (body) {
-        handleSensorDiscovery(req, res, body);
+        handlers.handleSensorDiscovery(req, res, body);
       });
     } else {
       jsonResponse(res, 405, { error: 'Method not allowed' });
     }
   } else if (urlPath === '/api/push/subscribe' && req.method === 'POST') {
     readBody(req, function (body) {
-      handlePushSubscribe(req, res, body);
+      handlers.handlePushSubscribe(req, res, body);
     });
   } else if (urlPath === '/api/push/unsubscribe' && req.method === 'POST') {
     readBody(req, function (body) {
-      handlePushUnsubscribe(req, res, body);
+      handlers.handlePushUnsubscribe(req, res, body);
     });
   } else if (urlPath === '/api/push/subscription' && req.method === 'POST') {
     readBody(req, function (body) {
-      handlePushGetSubscription(req, res, body);
+      handlers.handlePushGetSubscription(req, res, body);
     });
   } else if (urlPath === '/api/push/test' && req.method === 'POST') {
     readBody(req, function (body) {
-      handlePushTest(req, res, body);
+      handlers.handlePushTest(req, res, body);
     });
   } else if (urlPath === '/api/history') {
-    handleHistoryApi(req, res);
+    handlers.handleHistoryApi(req, res);
   } else if (urlPath === '/api/events') {
-    handleEventsApi(req, res);
+    handlers.handleEventsApi(req, res);
   } else if (urlPath === '/api/watchdog/state' && req.method === 'GET') {
     anomalyManager.getState().then(function (state) {
       jsonResponse(res, 200, state);
@@ -649,200 +387,11 @@ var server = http.createServer(function (req, res) {
   }
 });
 
-// ── WebSocket command handling ──
-
-var VALID_RELAYS = ['vi_btm', 'vi_top', 'vi_coll', 'vo_coll', 'vo_rad', 'vo_tank', 'v_air', 'pump', 'fan'];
-var overrideTtlTimer = null;
-
-function wsSend(ws, msg) {
-  if (ws.readyState === 1) ws.send(JSON.stringify(msg));
-}
-
-function handleWsCommand(ws, data) {
-  var msg;
-  try {
-    msg = JSON.parse(data.toString());
-  } catch (e) {
-    return;
-  }
-  if (!msg || !msg.type) return;
-
-  // Read-only sessions cannot mutate device state via the websocket.
-  if (ws._role && ws._role !== 'admin') {
-    wsSend(ws, { type: 'override-error', message: 'Admin role required' });
-    return;
-  }
-
-  if (msg.type === 'override-enter') {
-    handleOverrideEnter(ws, msg);
-  } else if (msg.type === 'override-exit') {
-    handleOverrideExit(ws);
-  } else if (msg.type === 'override-update') {
-    handleOverrideUpdate(ws, msg);
-  } else if (msg.type === 'override-set-mode') {
-    handleOverrideSetMode(ws, msg);
-  } else if (msg.type === 'relay-command') {
-    handleRelayCommand(ws, msg);
-  }
-}
-
-function handleOverrideEnter(ws, msg) {
-  var cfg = deviceConfig.getConfig();
-  if (!cfg.ce) {
-    wsSend(ws, { type: 'override-error', message: 'Controls not enabled' });
-    return;
-  }
-
-  // `fm` is REQUIRED when entering override (2026-04-21 hard-override
-  // semantics): automation is fully suspended for the duration, so the
-  // user must pick a concrete mode. The old "Automatic" state (fm=null
-  // while mo.a=true) is gone.
-  var VALID_MODES = ['I', 'SC', 'GH', 'AD', 'EH'];
-  var fm = msg.forcedMode;
-  if (typeof fm !== 'string' || VALID_MODES.indexOf(fm) === -1) {
-    wsSend(ws, { type: 'override-error', message: 'forcedMode required: one of I,SC,GH,AD,EH' });
-    return;
-  }
-
-  var ttl = Math.max(60, Math.min(3600, parseInt(msg.ttl, 10) || 300));
-  var ex = Math.floor(Date.now() / 1000) + ttl;
-
-  deviceConfig.updateConfig({ mo: { a: true, ex: ex, fm: fm } }, function (err, updated) {
-    if (err) {
-      wsSend(ws, { type: 'override-error', message: err.message });
-      return;
-    }
-    mqttBridge.publishConfig(updated);
-    wsSend(ws, { type: 'override-ack', active: true, expiresAt: ex, forcedMode: fm });
-
-    // Secondary server-side TTL tracking
-    clearOverrideTtlTimer();
-    overrideTtlTimer = setTimeout(function () {
-      overrideTtlTimer = null;
-      var current = deviceConfig.getConfig();
-      if (current.mo && current.mo.a) {
-        deviceConfig.updateConfig({ mo: null }, function (err2, cleared) {
-          if (!err2) mqttBridge.publishConfig(cleared);
-        });
-      }
-    }, ttl * 1000);
-  });
-}
-
-function handleOverrideExit(ws) {
-  clearOverrideTtlTimer();
-  deviceConfig.updateConfig({ mo: null }, function (err, updated) {
-    if (err) {
-      wsSend(ws, { type: 'override-error', message: err.message });
-      return;
-    }
-    mqttBridge.publishConfig(updated);
-    wsSend(ws, { type: 'override-ack', active: false, forcedMode: null });
-  });
-}
-
-function handleOverrideUpdate(ws, msg) {
-  var cfg = deviceConfig.getConfig();
-  if (!cfg.mo || !cfg.mo.a) {
-    wsSend(ws, { type: 'override-error', message: 'Override not active' });
-    return;
-  }
-
-  var ttl = Math.max(60, Math.min(3600, parseInt(msg.ttl, 10) || 300));
-  var ex = Math.floor(Date.now() / 1000) + ttl;
-
-  var newMo = { a: cfg.mo.a, ex: ex, fm: cfg.mo.fm };
-  deviceConfig.updateConfig({ mo: newMo }, function (err, updated) {
-    if (err) {
-      wsSend(ws, { type: 'override-error', message: err.message });
-      return;
-    }
-    mqttBridge.publishConfig(updated);
-    wsSend(ws, { type: 'override-ack', active: true, expiresAt: ex, forcedMode: (updated.mo && updated.mo.fm) || null });
-
-    // Reset secondary TTL timer
-    clearOverrideTtlTimer();
-    overrideTtlTimer = setTimeout(function () {
-      overrideTtlTimer = null;
-      var current = deviceConfig.getConfig();
-      if (current.mo && current.mo.a) {
-        deviceConfig.updateConfig({ mo: null }, function (err2, cleared) {
-          if (!err2) mqttBridge.publishConfig(cleared);
-        });
-      }
-    }, ttl * 1000);
-  });
-}
-
-function handleOverrideSetMode(ws, msg) {
-  var cfg = deviceConfig.getConfig();
-  if (!cfg.mo || !cfg.mo.a) {
-    wsSend(ws, { type: 'override-error', message: 'Override not active' });
-    return;
-  }
-
-  var mode = msg.mode;
-  var VALID_MODES = ['I', 'SC', 'GH', 'AD', 'EH'];
-  // With hard override, `fm` is required while active. Null/omit is no
-  // longer a legal state — server rejects it. If the user wants
-  // automation back, they must exit override.
-  if (typeof mode !== 'string' || VALID_MODES.indexOf(mode) === -1) {
-    wsSend(ws, { type: 'override-error', message: 'mode required: one of I,SC,GH,AD,EH' });
-    return;
-  }
-  if (cfg.wb && cfg.wb[mode] && cfg.wb[mode] > Math.floor(Date.now() / 1000)) {
-    wsSend(ws, { type: 'override-error', message: 'Mode banned' });
-    return;
-  }
-
-  var newMo = { a: cfg.mo.a, ex: cfg.mo.ex, fm: mode };
-
-  deviceConfig.updateConfig({ mo: newMo }, function (err, updated) {
-    if (err) {
-      wsSend(ws, { type: 'override-error', message: err.message });
-      return;
-    }
-    mqttBridge.publishConfig(updated);
-    wsSend(ws, {
-      type: 'override-ack',
-      active: true,
-      expiresAt: updated.mo.ex,
-      forcedMode: updated.mo.fm || null,
-    });
-  });
-}
-
-function handleRelayCommand(ws, msg) {
-  var cfg = deviceConfig.getConfig();
-  if (!cfg.mo || !cfg.mo.a) {
-    wsSend(ws, { type: 'override-error', message: 'Override not active' });
-    return;
-  }
-  var now = Math.floor(Date.now() / 1000);
-  if (cfg.mo.ex <= now) {
-    wsSend(ws, { type: 'override-error', message: 'Override expired' });
-    return;
-  }
-  if (VALID_RELAYS.indexOf(msg.relay) < 0) {
-    wsSend(ws, { type: 'override-error', message: 'Unknown relay: ' + msg.relay });
-    return;
-  }
-  mqttBridge.publishRelayCommand(msg.relay, !!msg.on);
-}
-
-function clearOverrideTtlTimer() {
-  if (overrideTtlTimer) {
-    clearTimeout(overrideTtlTimer);
-    overrideTtlTimer = null;
-  }
-}
-
-// ── WebSocket server ──
+// ── WebSocket server + broadcast ──
+// anomaly-manager calls broadcastToWebSockets to push watchdog-state.
 
 var wsServer = null;
 
-// Broadcast a message to all connected WebSocket clients. Used by the
-// anomaly-manager to push watchdog-state updates to the UI.
 function broadcastToWebSockets(msg) {
   if (!wsServer) return;
   var str = JSON.stringify(msg);
@@ -906,46 +455,19 @@ function initWebSocket() {
 
 // ── Startup ──
 
-function getNetworkAddress() {
-  var interfaces = os.networkInterfaces();
-  for (var name in interfaces) {
-    var addrs = interfaces[name];
-    for (var i = 0; i < addrs.length; i++) {
-      if (addrs[i].family === 'IPv4' && !addrs[i].internal) {
-        return addrs[i].address;
-      }
-    }
-  }
-  return null;
-}
-
-function printBanner(port, networkIp) {
-  var local = 'http://localhost:' + port;
-  var network = networkIp ? 'http://' + networkIp + ':' + port : null;
-
-  var lines = ['', '   Serving!', '', '   - Local:    ' + local];
-  if (network) lines.push('   - Network:  ' + network);
-  lines.push('');
-
-  var maxLen = 0;
-  for (var i = 0; i < lines.length; i++) {
-    if (lines[i].length > maxLen) maxLen = lines[i].length;
-  }
-  var width = maxLen + 4;
-  console.log('');
-  console.log('   \u250c' + '\u2500'.repeat(width) + '\u2510');
-  for (var j = 0; j < lines.length; j++) {
-    var padded = lines[j] + ' '.repeat(width - lines[j].length);
-    console.log('   \u2502' + padded + '\u2502');
-  }
-  console.log('   \u2514' + '\u2500'.repeat(width) + '\u2518');
-  console.log('');
-}
 
 function initServices(callback) {
   // Resolve DATABASE_URL from env or S3
   var dbModule = require('./lib/db');
   dbModule.resolveUrl(function (err, url) {
+    function finish() {
+      handlers = createHandlers({
+        db: db,
+        authMiddleware: authMiddleware,
+        broadcastToWebSockets: broadcastToWebSockets,
+      });
+      callback();
+    }
     if (url) {
       db = dbModule;
       db.initSchema(function (schemaErr) {
@@ -957,54 +479,25 @@ function initServices(callback) {
           db.startMaintenance();
         }
         initAnomalyManager();
-        callback();
+        finish();
       });
     } else {
       log.info('DATABASE_URL not found (checked env and S3) — history features disabled');
       initAnomalyManager();
-      callback();
+      finish();
     }
   });
 }
 
-// Initialize the anomaly-manager with its dependencies. Called after DB
-// init (or after confirming no DB is available) so the history storage
-// can pick the right backend (Postgres when db is present, ring buffer
-// fallback otherwise). Also applies the watchdog_events schema when a
-// database connection is available.
 function initAnomalyManager() {
-  // Apply the watchdog_events schema if we have a database connection.
-  if (db && typeof db.getPool === 'function') {
-    try {
-      var pool = db.getPool();
-      var sqlPath = path.join(__dirname, 'db', 'watchdog-events-schema.sql');
-      var schemaSql = fs.readFileSync(sqlPath, 'utf8');
-      pool.query(schemaSql, [], function (schemaErr) {
-        if (schemaErr) {
-          log.warn('watchdog schema init failed', { error: schemaErr.message });
-        } else {
-          log.info('watchdog_events schema ready');
-        }
-      });
-    } catch (e) {
-      log.warn('failed to apply watchdog schema', { error: e.message });
-    }
-  }
-
-  var wdHistoryDb = (db && typeof db.getPool === 'function') ? db.getPool() : null;
-  var watchdogHistory = createWatchdogHistory({
-    db: wdHistoryDb,
-    log: log
-  });
-  anomalyManager.init({
-    history: watchdogHistory,
+  anomalyManager.bootstrap({
+    db: db,
     push: push,
     wsBroadcast: broadcastToWebSockets,
     mqttBridge: mqttBridge,
     deviceConfig: deviceConfig,
-    log: log
+    log: log,
   });
-  log.info('anomaly-manager initialized', { backend: wdHistoryDb ? 'postgres' : 'ring-buffer' });
 }
 
 function startMqttBridge() {
@@ -1094,8 +587,8 @@ if (require.main === module) {
 }
 
 // ── Test-only exports ──
-// Exported so unit tests can exercise WS command handlers without starting the server.
+// WS command handlers moved to ./lib/ws-command-handlers.js; tests
+// require that module directly now.
 module.exports = {
   handleWsCommand: handleWsCommand,
-  wsSend: wsSend,
 };
